+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 - app - deploy jslint as chrome-extension.
 - jslint - add `for...of` syntax support.
 - jslint - add html and css linting back into jslint.
-- jslint - add new warning if const/let/var statements are not declared at top of function-scope.
 - jslint - add new warning if const/let/var statements are not sorted.
 - jslint - include explicit commonjs (jslint.cjs) and es-module (jslint.mjs) variants of jslint.
 - jslint - remove directive "eval" (use line-specific ignore-directive "//jslint-quiet" instead).
@@ -18,6 +17,7 @@
 ## v2021.6.14-beta
 - bugfix - fix global_list being ignored by jslint.
 - ci - migrate ci-scripts from cjs to esm.
+- jslint - add new warning if const/let/var statements are not declared at top of function-scope.
 - website - invalidate url-cache with each deployment.
 - website - replace .png logo with .svg logo.
 - website - replace current-editor with CodeMirror-editor and change programming-font-family from `Programma` to `consolas, menlo, monospace`.